wip(data-aggregation): isolate treatment-day-count calculation
Due to Prisma's limitation, currently they doesn't support feature
like group by day, which we really need to make a summary dashboard.
So, we've figured out 2 solution.

1. Using raw query
2. Isolate treatment-day-count then store in db

The first solution is quite make a raw-query to never reflect our
prisma's schema. But, the last one still consist our schema. And
we would to go for the last.

To isolate the calculate then store in db, we need a cronjob to
execute our calculation daily. That why we implemented a webhook's
service. We planed to use GitHub Actions's feature to do the job.